feat: remove legacy map; add rule for vec'zd map
